Code cleanup, structure reorganization, old files removed.
