--- server/ro.plo
@@ -1,504 +1,0 @@
-S 0
-L -200 1
-R 400 1
-L 400 1
-R -600 1
-L -600 1
-R 800 1
-L 800 1
-R -1000 1
-L -1000 1
-R 1200 1
-L 1200 1
-R -1400 1
-L -1400 1
-R 1600 1
-L 1600 1
-R -1800 1
-L -1800 1
-R 2000 1
-L 2000 1
-R -2200 1
-L -2200 1
-R 2400 1
-L 2400 1
-R -2600 1
-L -2600 1
-R 2800 1
-L 2800 1
-R -3000 1
-L -3000 1
-R 3200 1
-L 3200 1
-R -3400 1
-L -3400 1
-R 3600 1
-L 3600 1
-R -3800 1
-L -3800 1
-R 4000 1
-L 4000 1
-R -4200 1
-L -4200 1
-R 4400 1
-L 4400 1
-R -4600 1
-L -4600 1
-R 4800 1
-L 4800 1
-R -5000 1
-L -5000 1
-R 5200 1
-L 5200 1
-R -5400 1
-L -5400 1
-R 5600 1
-L 5600 1
-R -5800 1
-L -5800 1
-R 6000 1
-L 6000 1
-R -6200 1
-L -6200 1
-R 6400 1
-L 6400 1
-R -6600 1
-L -6600 1
-R 6800 1
-L 6800 1
-R -7000 1
-L -7000 1
-R 7200 1
-L 7200 1
-R -7400 1
-L -7400 1
-R 7600 1
-L 7600 1
-R -7800 1
-L -7800 1
-R 8000 1
-L 8000 1
-R -8200 1
-L -8200 1
-R 8400 1
-L 8400 1
-R -8600 1
-L -8600 1
-R 8800 1
-L 8800 1
-R -9000 1
-L -9000 1
-R 9200 1
-L 9200 1
-R -9400 1
-L -9400 1
-R 9600 1
-L 9600 1
-R -9800 1
-L -9800 1
-R 10000 1
-L 10000 1
-R -10200 1
-L -10200 1
-R 10400 1
-L 10400 1
-R -10600 1
-L -10600 1
-R 10800 1
-L 10800 1
-R -11000 1
-L -11000 1
-R 11200 1
-L 11200 1
-R -11400 1
-L -11400 1
-R 11600 1
-L 11600 1
-R -11800 1
-L -11800 1
-R 12000 1
-L 12000 1
-R -12200 1
-L -12200 1
-R 12400 1
-L 12400 1
-R -12600 1
-L -12600 1
-R 12800 1
-L 12800 1
-R -13000 1
-L -13000 1
-R 13200 1
-L 13200 1
-R -13400 1
-L -13400 1
-R 13600 1
-L 13600 1
-R -13800 1
-L -13800 1
-R 14000 1
-L 14000 1
-R -14200 1
-L -14200 1
-R 14400 1
-L 14400 1
-R -14600 1
-L -14600 1
-R 14800 1
-L 14800 1
-R -15000 1
-L -15000 1
-R 15200 1
-L 15200 1
-R -15400 1
-L -15400 1
-R 15600 1
-L 15600 1
-R -15800 1
-L -15800 1
-R 16000 1
-L 16000 1
-R -16200 1
-L -16200 1
-R 16400 1
-L 16400 1
-R -16600 1
-L -16600 1
-R 16800 1
-L 16800 1
-R -17000 1
-L -17000 1
-R 17200 1
-L 17200 1
-R -17400 1
-L -17400 1
-R 17600 1
-L 17600 1
-R -17800 1
-L -17800 1
-R 18000 1
-L 18000 1
-R -18200 1
-L -18200 1
-R 18400 1
-L 18400 1
-R -18600 1
-L -18600 1
-R 18800 1
-L 18800 1
-R -19000 1
-L -19000 1
-R 19200 1
-L 19200 1
-R -19400 1
-L -19400 1
-R 19600 1
-L 19600 1
-R -19800 1
-L -19800 1
-R 20000 1
-L 20000 1
-R -20200 1
-L -20200 1
-R 20400 1
-L 20400 1
-R -20600 1
-L -20600 1
-R 20800 1
-L 20800 1
-R -21000 1
-L -21000 1
-R 21200 1
-L 21200 1
-R -21400 1
-L -21400 1
-R 21600 1
-L 21600 1
-R -21800 1
-L -21800 1
-R 22000 1
-L 22000 1
-R -22200 1
-L -22200 1
-R 22400 1
-L 22400 1
-R -22600 1
-L -22600 1
-R 22800 1
-L 22800 1
-R -23000 1
-L -23000 1
-R 23200 1
-L 23200 1
-R -23400 1
-L -23400 1
-R 23600 1
-L 23600 1
-R -23800 1
-L -23800 1
-R 24000 1
-L 24000 1
-R -24200 1
-L -24200 1
-R 24400 1
-L 24400 1
-R -24600 1
-L -24600 1
-R 24800 1
-L 24800 1
-R -25000 1
-L -25000 1
-R 25200 1
-L 25200 1
-R -25400 1
-L -25400 1
-R 25600 1
-L 25600 1
-R -25800 1
-L -25800 1
-R 26000 1
-L 26000 1
-R -26200 1
-L -26200 1
-R 26400 1
-L 26400 1
-R -26600 1
-L -26600 1
-R 26800 1
-L 26800 1
-R -27000 1
-L -27000 1
-R 27200 1
-L 27200 1
-R -27400 1
-L -27400 1
-R 27600 1
-L 27600 1
-R -27800 1
-L -27800 1
-R 28000 1
-L 28000 1
-R -28200 1
-L -28200 1
-R 28400 1
-L 28400 1
-R -28600 1
-L -28600 1
-R 28800 1
-L 28800 1
-R -29000 1
-L -29000 1
-R 29200 1
-L 29200 1
-R -29400 1
-L -29400 1
-R 29600 1
-L 29600 1
-R -29800 1
-L -29800 1
-R 30000 1
-L 30000 1
-R -30200 1
-L -30200 1
-R 30400 1
-L 30400 1
-R -30600 1
-L -30600 1
-R 30800 1
-L 30800 1
-R -31000 1
-L -31000 1
-R 31200 1
-L 31200 1
-R -31400 1
-L -31400 1
-R 31600 1
-L 31600 1
-R -31800 1
-L -31800 1
-R 32000 1
-L 32000 1
-R -32200 1
-L -32200 1
-R 32400 1
-L 32400 1
-R -32600 1
-L -32600 1
-R 32800 1
-L 32800 1
-R -33000 1
-L -33000 1
-R 33200 1
-L 33200 1
-R -33400 1
-L -33400 1
-R 33600 1
-L 33600 1
-R -33800 1
-L -33800 1
-R 34000 1
-L 34000 1
-R -34200 1
-L -34200 1
-R 34400 1
-L 34400 1
-R -34600 1
-L -34600 1
-R 34800 1
-L 34800 1
-R -35000 1
-L -35000 1
-R 35200 1
-L 35200 1
-R -35400 1
-L -35400 1
-R 35600 1
-L 35600 1
-R -35800 1
-L -35800 1
-R 36000 1
-L 36000 1
-R -36200 1
-L -36200 1
-R 36400 1
-L 36400 1
-R -36600 1
-L -36600 1
-R 36800 1
-L 36800 1
-R -37000 1
-L -37000 1
-R 37200 1
-L 37200 1
-R -37400 1
-L -37400 1
-R 37600 1
-L 37600 1
-R -37800 1
-L -37800 1
-R 38000 1
-L 38000 1
-R -38200 1
-L -38200 1
-R 38400 1
-L 38400 1
-R -38600 1
-L -38600 1
-R 38800 1
-L 38800 1
-R -39000 1
-L -39000 1
-R 39200 1
-L 39200 1
-R -39400 1
-L -39400 1
-R 39600 1
-L 39600 1
-R -39800 1
-L -39800 1
-R 40000 1
-L 40000 1
-R -40200 1
-L -40200 1
-R 40400 1
-L 40400 1
-R -40600 1
-L -40600 1
-R 40800 1
-L 40800 1
-R -41000 1
-L -41000 1
-R 41200 1
-L 41200 1
-R -41400 1
-L -41400 1
-R 41600 1
-L 41600 1
-R -41800 1
-L -41800 1
-R 42000 1
-L 42000 1
-R -42200 1
-L -42200 1
-R 42400 1
-L 42400 1
-R -42600 1
-L -42600 1
-R 42800 1
-L 42800 1
-R -43000 1
-L -43000 1
-R 43200 1
-L 43200 1
-R -43400 1
-L -43400 1
-R 43600 1
-L 43600 1
-R -43800 1
-L -43800 1
-R 44000 1
-L 44000 1
-R -44200 1
-L -44200 1
-R 44400 1
-L 44400 1
-R -44600 1
-L -44600 1
-R 44800 1
-L 44800 1
-R -45000 1
-L -45000 1
-R 45200 1
-L 45200 1
-R -45400 1
-L -45400 1
-R 45600 1
-L 45600 1
-R -45800 1
-L -45800 1
-R 46000 1
-L 46000 1
-R -46200 1
-L -46200 1
-R 46400 1
-L 46400 1
-R -46600 1
-L -46600 1
-R 46800 1
-L 46800 1
-R -47000 1
-L -47000 1
-R 47200 1
-L 47200 1
-R -47400 1
-L -47400 1
-R 47600 1
-L 47600 1
-R -47800 1
-L -47800 1
-R 48000 1
-L 48000 1
-R -48200 1
-L -48200 1
-R 48400 1
-L 48400 1
-R -48600 1
-L -48600 1
-R 48800 1
-L 48800 1
-R -49000 1
-L -49000 1
-R 49200 1
-L 49200 1
-R -49400 1
-L -49400 1
-R 49600 1
-L 49600 1
-R -49800 1
-L -49800 1
-R 50000 1
-L 50000 1
-R -50200 1
-S 1
-P 10
-
